refactor : seatRock 제외 + test 추가
